change js structure to modular
