+++ app.js
@@ -1,0 +1,5 @@
+import { toggleCardBookmark } from "./modules/bookmark.js";
+import { navigation } from "./modules/navigation.js";
+
+toggleCardBookmark();
+navigation();
