+++ index.html
@@ -356,6 +356,6 @@
         </ul>
       </nav>
     </footer>
-    <script type="text/javascript" src="js/app.js"></script>
+    <script type="module" src="/app.js"></script>
   </body>
 </html>
